+++ packages/appcd/src/common.js
@@ -107,7 +107,7 @@ export function loadConfig(argv) {
  * for the Node debugger.
  * @returns {Promise}
  */
-export function startServer({ cfg, argv }) {
+export async function startServer({ cfg, argv }) {
 	const { config, configFile, debug, debugInspect } = argv;
 	const args = [];
 	const detached = debug || debugInspect ? false : cfg.get('server.daemonize');
@@ -137,32 +137,32 @@ export function startServer({ cfg, argv }) {
 	}
 	process.env.FORCE_COLOR = 1;
 
-	return Promise.resolve()
-		.then(() => import('appcd-nodejs'))
-		.then(({ generateV8MemoryArgument, spawnNode }) => {
-			// check if we should use the core's required Node.js version
-			if (cfg.get('core.enforceNodeVersion') !== false) {
-				if (!nodeVer) {
-					throw new Error(`Invalid Node.js engine version from appcd-core package.json: ${nodeVer}`);
-				}
+	try {
+		const { generateV8MemoryArgument, spawnNode } = await import('appcd-nodejs');
+		let child;
 
-				return spawnNode({
-					args,
-
-					// On macOS (and probably Linux), a detached process doesn't stick around to see
-					// if the executable exited with an error, so we must not detach the process,
-					// but rather disconnect it once the daemon is booted. On Windows, we have to
-					// detach it to keep the process running and for some lucky reason, Node sticks
-					// around to see if the executable errors.
-					detached: process.platform === 'win32' ? detached : false,
-
-					nodeHome: expandPath(cfg.get('home'), 'node'),
-					stdio,
-					v8mem,
-					version: nodeVer
-				});
+		// check if we should use the core's required Node.js version
+		if (cfg.get('core.enforceNodeVersion') !== false) {
+			if (!nodeVer) {
+				throw new Error(`Invalid Node.js engine version from appcd-core package.json: ${nodeVer}`);
 			}
 
+			child = await spawnNode({
+				args,
+
+				// On macOS (and probably Linux), a detached process doesn't stick around to see
+				// if the executable exited with an error, so we must not detach the process,
+				// but rather disconnect it once the daemon is booted. On Windows, we have to
+				// detach it to keep the process running and for some lucky reason, Node sticks
+				// around to see if the executable errors.
+				detached: process.platform === 'win32' ? detached : false,
+
+				nodeHome: expandPath(cfg.get('home'), 'node'),
+				stdio,
+				v8mem,
+				version: nodeVer
+			});
+		} else {
 			// using the current Node.js version which may be incompatible with the core
 
 			if (v8mem) {
@@ -172,9 +172,16 @@ export function startServer({ cfg, argv }) {
 				}
 			}
 
-			return spawn(process.execPath, args, { stdio });
-		})
-		.then(child => new Promise((resolve, reject) => {
+			child = spawn(process.execPath, args, { stdio });
+		}
+
+		if (debug || debugInspect) {
+			process
+				.on('SIGINT', () => child.kill('SIGINT'))
+				.on('SIGTERM', () => child.kill('SIGTERM'));
+		}
+
+		await new Promise((resolve, reject) => {
 			child.on('message', msg => {
 				if (msg === 'booted') {
 					if (detached) {
@@ -205,11 +212,11 @@ export function startServer({ cfg, argv }) {
 					resolve();
 				}
 			});
-		}))
-		.catch(err => {
-			log(err);
-			throw err;
 		});
+	} catch (err) {
+		log(err);
+		throw err;
+	}
 }
 
 /**