+++ packages/appcd-plugin/src/internal-plugin.js
@@ -1,11 +1,7 @@
-import appcdLogger from 'appcd-logger';
 import Dispatcher from 'appcd-dispatcher';
 import PluginBase from './plugin-base';
 
 import { AppcdError, codes } from 'appcd-response';
-
-const { log } = appcdLogger('appcd:plugin:internal');
-const { highlight } = appcdLogger.styles;
 
 /**
  * Internal plugin implementation logic.
@@ -38,17 +34,6 @@ export default class InternalPlugin extends PluginBase {
 	 */
 	async onStart() {
 		try {
-			if (this.plugin.configFile) {
-				log('Loading plugin config file:', highlight(this.plugin.configFile));
-				await Dispatcher.call('/appcd/config', {
-					data: {
-						action:   'load',
-						file:     this.plugin.configFile,
-						override: false
-					}
-				});
-			}
-
 			await this.activate();
 		} catch (err) {
 			this.info.error = err.message;